+++ examples/scripts/battery_parameterisation/bayesian_feature_fitting.py
@@ -11,9 +11,21 @@ parameter_values = pybamm.ParameterValues("Chen2020")
 original_D_n = parameter_values["Negative particle diffusivity [m2.s-1]"]
 original_D_p = parameter_values["Positive particle diffusivity [m2.s-1]"]
 
-# Put empty Parameter slots as placeholders
-parameter_values["Negative particle diffusivity [m2.s-1]"] = pybop.Parameter()
-parameter_values["Positive particle diffusivity [m2.s-1]"] = pybop.Parameter()
+# Set multivariate parameters (defined in model space)
+distribution = pybop.MultivariateLogNormal(
+    mean_log_x=[np.log(original_D_n), np.log(original_D_p)],
+    covariance_log_x=[[np.log(2), 0.0], [0.0, np.log(2)]],
+)
+parameter_values["Negative particle diffusivity [m2.s-1]"] = pybop.Parameter(
+    initial_value=0.9 * original_D_n,
+    transformation=pybop.LogTransformation(),
+    distribution=pybop.MarginalDistribution(distribution, 0),
+)
+parameter_values["Positive particle diffusivity [m2.s-1]"] = pybop.Parameter(
+    initial_value=1.1 * original_D_p,
+    transformation=pybop.LogTransformation(),
+    distribution=pybop.MarginalDistribution(distribution, 1),
+)
 
 # Set up simulator with custom settings
 submesh_types, var_pts, spatial_methods = spectral_mesh_pts_and_method(10, 10, 10)
@@ -52,26 +64,6 @@ synthetic_data = simulator.solve(
 )
 dataset = pybop.import_pybamm_solution(synthetic_data)
 
-# Override the forced univariate Parameters
-simulator.parameters = pybop.MultivariateParameters(
-    {
-        "Negative particle diffusivity [m2.s-1]": pybop.Parameter(
-            initial_value=0.9 * original_D_n,
-            bounds=[original_D_n / 2, original_D_n * 2],
-            transformation=pybop.LogTransformation(),
-        ),
-        "Positive particle diffusivity [m2.s-1]": pybop.Parameter(
-            initial_value=1.1 * original_D_p,
-            bounds=[original_D_p / 2, original_D_p * 2],
-            transformation=pybop.LogTransformation(),
-        ),
-    },
-    distribution=pybop.MultivariateGaussian(
-        [np.log(original_D_n), np.log(original_D_p)],
-        [[np.log(2), 0.0], [0.0, np.log(2)]],
-    ),
-)
-
 ICI_cost = pybop.SquareRootFeatureDistance(
     dataset["Time [s]"],
     dataset["Voltage [V]"],
@@ -92,9 +84,6 @@ if __name__ == "__main__":
     GITT_problem = pybop.Problem(simulator, GITT_cost)
     problem = pybop.MetaProblem(ICI_problem, GITT_problem)
 
-    # Copy the MultivariateParameters to the meta-problem
-    problem.parameters = simulator.parameters
-
     # Set up and run the optimiser, increase the number of iterations
     # and samples to improve accuracy
     options = pybop.EPBOLFIOptions(
@@ -105,6 +94,16 @@ if __name__ == "__main__":
         bolfi_posterior_effective_sample_size=10,
         posterior_gelman_rubin_threshold=1.2,
         verbose=True,
+        model_parameter_boundaries={
+            "Negative particle diffusivity [m2.s-1]": [
+                original_D_n / 2,
+                original_D_n * 2,
+            ],
+            "Positive particle diffusivity [m2.s-1]": [
+                original_D_p / 2,
+                original_D_p * 2,
+            ],
+        },
     )
     optim = pybop.EP_BOLFI(problem, options=options)
     result = optim.run()