+++ tests/unit/test_likelihoods.py
@@ -63,19 +63,19 @@ class TestLikelihoods:
         with pytest.raises(NotImplementedError):
             likelihood(np.array([0.5]))
 
-    def test_likelihood_check_sigma0(self, dataset):
+    def test_likelihood_check_sigma(self, dataset):
         with pytest.raises(ValueError, match="Sigma0 must be positive"):
-            pybop.GaussianLogLikelihoodKnownSigma(dataset, sigma0=None)
+            pybop.GaussianLogLikelihoodKnownSigma(dataset, sigma=None)
 
         likelihood = pybop.GaussianLogLikelihoodKnownSigma(dataset, 0.1)
-        likelihood.set_sigma0(0.2)
+        likelihood.set_sigma(0.2)
         assert likelihood.sigma2 == 0.2**2.0
 
         with pytest.raises(
             ValueError,
-            match=r"sigma0 must be either a scalar value",
+            match=r"sigma must be either a scalar value",
         ):
-            pybop.GaussianLogLikelihoodKnownSigma(dataset, sigma0=[0.2, 0.3])
+            pybop.GaussianLogLikelihoodKnownSigma(dataset, sigma=[0.2, 0.3])
 
     @pytest.mark.parametrize(
         "target",
@@ -83,7 +83,7 @@ class TestLikelihoods:
     )
     def test_gaussian_log_likelihood_known_sigma(self, simulator, dataset, target):
         likelihood = pybop.GaussianLogLikelihoodKnownSigma(
-            dataset, target=target, sigma0=np.array([0.01])
+            dataset, target=target, sigma=np.array([0.01])
         )
         problem = pybop.Problem(simulator, likelihood)
         result = problem.evaluate([0.5]).values
@@ -96,7 +96,7 @@ class TestLikelihoods:
         assert grad_likelihood >= 0
 
     def test_gaussian_log_likelihood(self, simulator, dataset):
-        likelihood = pybop.GaussianLogLikelihood(dataset, sigma0=0.01)
+        likelihood = pybop.GaussianLogLikelihood(dataset, sigma=0.01)
         problem = pybop.Problem(simulator, likelihood)
         result = problem.evaluate(np.array([0.8, 0.02])).values
         grad_result, grad_likelihood = problem.evaluate(
@@ -111,14 +111,14 @@ class TestLikelihoods:
 
         # Test construction with sigma as a Parameter
         sigma = pybop.Parameter(stats.uniform(loc=0.4, scale=0.6 - 0.4))
-        likelihood = pybop.GaussianLogLikelihood(dataset, sigma0=sigma)
+        likelihood = pybop.GaussianLogLikelihood(dataset, sigma=sigma)
 
         # Test invalid sigma
         with pytest.raises(
             TypeError,
-            match=r"Expected sigma0 to contain Parameter objects or numeric values.",
+            match=r"Expected sigma to contain Parameter objects or numeric values.",
         ):
-            likelihood = pybop.GaussianLogLikelihood(dataset, sigma0="Invalid string")
+            likelihood = pybop.GaussianLogLikelihood(dataset, sigma="Invalid string")
             pybop.Problem(simulator, likelihood)
 
     def test_gaussian_log_likelihood_dsigma_scale(self, dataset):
@@ -146,7 +146,7 @@ class TestLikelihoods:
         self, simulator, dataset
     ):
         likelihood = pybop.GaussianLogLikelihoodKnownSigma(
-            dataset, sigma0=np.array([0.2])
+            dataset, sigma=np.array([0.2])
         )
         problem = pybop.Problem(simulator, likelihood)
         assert (